move bundle & manifest related functions to bundle
