+++ rsconnect/api.py
@@ -8,14 +8,14 @@ from _ssl import SSLError
 import re
 from warnings import warn
 from six import text_type
-from rsconnect.instrumentation import (
+from .bundle import (
     _warn_if_environment_directory,
     _warn_if_no_requirements_file,
     _warn_on_ignored_manifest,
     _warn_on_ignored_requirements,
     are_apis_supported_on_server,
-    fake_module_file_from_directory,
     get_python_env_info,
+    fake_module_file_from_directory,
     validate_entry_point,
     validate_extra_files,
 )
